+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Fix [#144]: remove configuration file as it creates more problems than solves.
+
 ## [v1.0.8]
 
 - Add: remove config message for CLI app.